+++ src/DotRecast.Recast/EdgeValues.cs
@@ -1,0 +1,8 @@
+﻿namespace DotRecast.Recast
+{
+    public static class EdgeValues
+    {
+        public const int EV_UNDEF = -1;
+        public const int EV_HULL = -2;
+    }
+}