+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Nothing
 
 ### Changed
-- Nothing
+- Changed to reuse samples and edges list in BuildPolyDetail()
 
 ### Removed
 - Removed RcVecUtils.Dot()